research/postgres/requirements.txt
pytest==7.1.2
Faker==14.1.0
psycopg2-binary==2.9.3